Added self-building raylib with JPG support
It patches itself on build. Linux build support is currently broken with the .sh files, unfortunately for now it's just a case of "work it out yourself", since the sh scripts are currently old and reference parts of the code that no longer exist. Also, changed gitignore to accomodate new changes, and zoombinis.7z to accomodate file structure.